Adding RAG Handling
