+++ tests/unittest/test_pr_rag_engine.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from pr_agent.algo.types import EDIT_TYPE
-from pr_agent.tools.pr_rag_engine import PRRagEngine
+from pr_agent.tools.pr_rag_engine import PRRAGEngine
 
 
 def test_documents():
@@ -70,7 +70,7 @@ def mock_git_provider():
     return provider
 
 def test_create_new_pr_index_with_diff_files(mock_rag_client, mock_git_provider):
-    engine = PRRagEngine("http://fake-url")
+    engine = PRRAGEngine("http://fake-url")
     engine.rag_client = mock_rag_client
 
     # Patch index_name creation
@@ -90,7 +90,7 @@ def test_create_new_pr_index_with_diff_files(mock_rag_client, mock_git_provider)
         assert args[1][0]["metadata"]["split_type"] == "code"
 
 def test_update_index_add_modify_delete(mock_rag_client, mock_git_provider):
-    engine = PRRagEngine("http://fake-url")
+    engine = PRRAGEngine("http://fake-url")
     engine.rag_client = mock_rag_client
 
     mock_git_provider.get_diff_files.return_value = [
@@ -121,7 +121,7 @@ def test_update_index_add_modify_delete(mock_rag_client, mock_git_provider):
 
 def test_create_new_base_index(mock_rag_client, mock_git_provider):
     mock_rag_client.list_indexes.return_value = []
-    engine = PRRagEngine("http://fake-url")
+    engine = PRRAGEngine("http://fake-url")
     engine.rag_client = mock_rag_client
     with patch.object(engine, '_get_git_provider', return_value=mock_git_provider):
         engine.create_base_branch_index("http://pr-url")
@@ -135,7 +135,7 @@ def test_create_new_base_index(mock_rag_client, mock_git_provider):
         assert args[1][0]["metadata"]["split_type"] == "code"
 
 def test_delete_pr_index(mock_rag_client, mock_git_provider):
-    engine = PRRagEngine("http://fake-url")
+    engine = PRRAGEngine("http://fake-url")
     engine.rag_client = mock_rag_client
 
     with patch.object(engine, '_get_git_provider', return_value=mock_git_provider):
